Extracting the play and pause methods into a useAudio hook
