+++ src/hooks/useAudio.js
@@ -1,0 +1,56 @@
+import React, { useState, useEffect } from "react";
+import { defaultSongImage } from "../utils/defaultImage";
+
+export default function useAudio(song, audio, setAudio, setAudioDetails) {
+  const [backgroundAudio, setBackgroundAudio] = useState(null);
+
+  useEffect(() => {
+    loadAudio();
+  }, []);
+
+  const songSource = song.songLocation;
+
+  const loadAudio = () => {
+    console.log("Background loading audio for song: " + song.name);
+    let loadingAudio = new Audio();
+    loadingAudio.src = songSource;
+    loadingAudio.load();
+    setBackgroundAudio(loadingAudio);
+  };
+
+  const handlePlay = () => {
+    if (audio.src !== songSource) {
+      pauseAndRewindCurrentAudio();
+      replaceWithNewAudioAndStartPlaying();
+    } else {
+      audio.play();
+    }
+    setAudioDetails({
+      isPlaying: true,
+      source: songSource,
+      name: song.name,
+      image: song.image || defaultSongImage,
+    });
+  };
+
+  const handlePause = () => {
+    audio.pause();
+    setAudioDetails((previous) => {
+      return { ...previous, isPlaying: false };
+    });
+  };
+
+  const pauseAndRewindCurrentAudio = () => {
+    audio.pause();
+    audio.currentTime = 0;
+  };
+
+  const replaceWithNewAudioAndStartPlaying = () => {
+    backgroundAudio.play();
+    backgroundAudio.volume = audio.volume;
+    backgroundAudio.loop = audio.loop;
+    setAudio(backgroundAudio);
+  };
+
+  return [handlePlay, handlePause];
+}
